Refactor: Move applied filters in a new component
